revision

+++ src/test/java/SignUpTest.java
@@ -18,18 +18,23 @@ import static RestApis.BaseUserTest.getAuthPj;
 import static org.junit.Assert.assertEquals;
 
 public class SignUpTest extends BaseTest {
+
     @Before
     @Override
     public void setup() {
-        RestAssured.baseURI = "https://stellarburgers.nomoreparties.site";
+        RestAssured.baseURI = URLS.siteURL;
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
+
+        // Создаем тестового пользователя
         pj = createUser();
 
+        // Инициализируем WebDriver и страницы
         try {
             driver = getWebDriver("chrome");
         } catch (MalformedURLException e) {
-            System.out.println("error" + e.toString());
+            System.out.println("Ошибка при инициализации WebDriver: " + e.toString());
         }
+
         acc = new Account(driver);
         home = new Home(driver);
         login = new Login(driver);
@@ -42,12 +47,15 @@ public class SignUpTest extends BaseTest {
     public void WrongPassTest() {
         home.accountButtonClick();
         signUp.signUpLinkClick();
-        pj.setPassword("passw");
+        pj.setPassword("passw"); // Не валидный пароль
         signUp.sendSignUpData(pj);
         signUp.clickSignUpButton();
+
+        // Проверяем сообщение об ошибке
         String s = signUp.getErrorPasswordMessage();
         assertEquals("Некорректный пароль", s);
     }
+
     @Test
     @DisplayName("Тест Успешная регистрация")
     public void RegistrationTest() {
@@ -59,18 +67,22 @@ public class SignUpTest extends BaseTest {
         login.clickLogInButton();
         home.accountButtonClick();
 
+        // Проверка данных пользователя
         PojoUser actual_pj = acc.getData();
-        assertEquals(pj.getName().toLowerCase(),actual_pj.getName().toLowerCase());
-        assertEquals(pj.getEmail().toLowerCase(),actual_pj.getEmail().toLowerCase());
+        assertEquals(pj.getName().toLowerCase(), actual_pj.getName().toLowerCase());
+        assertEquals(pj.getEmail().toLowerCase(), actual_pj.getEmail().toLowerCase());
 
         acc.logOutButtonClick();
-        BaseUserTest.deleteUser(pj);
-
     }
+
     @After
     @Override
     public void teardown() {
-        driver.quit();
+        if (pj != null && pj.getAccessToken() != null) {
+            BaseUserTest.deleteUser(pj); // Удаляем пользователя после теста
+        }
+        if (driver != null) {
+            driver.quit(); // Закрываем драйвер после теста
+        }
     }
-
 }
